Fix build image change race
The build image change controller was performing the following logic
once it determined a build was needed:

1. Bump the BuildConfig so the ResourceVersion would be updated, in
theory to prevent other image change controllers from being able to
trigger a build for the same BuildConfig revision.
2. Make an API call to instantiate a build (BuildGenerator)
3. The BuildGenerator would try to modify the BuildConfig's LastVersion
4. The BuildGenerator would try to create a Build
5. The image change controller would try to update the BuildConfig to
set the new LastTriggeredImageID value.

If another instance of an image change controller bumped the
BuildConfig's ResourceVersion before 3 above, you could get into an
infinite race loop where 1 controller would keep bumping the
ResourceVersion just before the BuildGenerator tried to update the
config's LastVersion.

This change makes it so the BuildGenerator is the only thing responsible
for both creating a Build and updating the BuildConfig. The Build is
created prior to updating the BuildConfig. If the Build creation fails,
it won't try to update the BuildConfig.

BuildRequest gains a new required field (Image) that the image change
controller sets to tell the BuildGenerator what value to set for
LastTriggeredImageID.